+++ src/main/java/com/bidnbuy/server/dto/BidUpdateDto.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 public class BidUpdateDto {
-    private Integer auctionId; // 어떤 경매가 갱신되었는지
+    private Long auctionId; // 어떤 경매가 갱신되었는지
     private Integer currentPrice; // 새로 갱신된 현재 최고가
     private Integer bidCount; // 새로 갱신된 입찰 횟수
     private Long lastBidderId; // 마지막 입찰자 ID (선택 사항)
